Dev (#9)
* update dabase schema

* modify data schema

* modify schema

* add favorite method

* add doc many to many

* add role authorized and middleware

* fix favorite list method

* add user/admin middleware

* add a common role/user check

* adjust api route of favorites

* adjust authorization to recipe and review controller

* add access check to ingredient update and destroy

* fix error call checkrole

* fix role check method call

* finish backend user profile api

* front end adjust for user profile

* adjust output of recipelist from be

* fix fe import issues

* fix fe import issues

* adjust user recipes

* implement user detail page

* add scss

* add useEffect

* unify current user output from be

* update a bit style of FORM

* adjust file structure

* recipe list and details

* add favorite, fix user detail

* finish user profile

* fix authentication of user controller

* clean codes

* adjust user factory, delete console

* add pictures

* add review count and favorite count

+++ backend/app/Http/Controllers/Api/IngredientController.php
@@ -92,6 +92,10 @@ class IngredientController extends Controller
     {
         // Update a specific ingredient by its ID
         try{
+            if(!$this->checkRole('admin')) {
+                return $this->sendError('Unauthorized', [], 401);
+            }
+
             $ingredient = Ingredient::findOrFail($id);
             $validatedData = $request->validate([
                 'name' => 'nullable|string|max:50',
@@ -119,6 +123,10 @@ class IngredientController extends Controller
     {
         // Delete a specific ingredient by its ID
         try {
+            if(!$this->checkRole('admin')) {
+                return $this->sendError('Unauthorized', [], 401);
+            }
+
             $ingredient = Ingredient::findOrFail($id);
             $ingredient->delete();
